@@ -7,8 +7,8 @@ describe('ng-shop App', () => {
     page = new NgShopPage();
   });
 
-  it('should display message saying app works', () => {
+  it('should display welcome message', () => {
     page.navigateTo();
-    expect(page.getParagraphText()).toEqual('app works!');
+    expect(page.getParagraphText()).toEqual('Welcome to ngs!');
   });
 });
