@@ -2,8 +2,18 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
-import { MaterialModule } from '@angular/material';
+import { HttpClientModule } from '@angular/common/http';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterModule } from '@angular/router';
+import {
+  MdButtonModule,
+  MdIconModule,
+  MdInputModule,
+  MdSelectModule,
+  MdTabsModule,
+  MdToolbarModule,
+  MdGridListModule
+} from '@angular/material';
 
 import { ProductService, ShoppingCartService } from './shared/services';
 import { routes } from './app.routing';
@@ -13,14 +23,6 @@ import { HomeComponent, ProductTileComponent } from './home';
 import { ProductComponent, ProductDetailsComponent } from './product';
 
 @NgModule({
-  imports: [
-    BrowserModule,
-    FormsModule,
-    HttpModule,
-    ReactiveFormsModule,
-    RouterModule.forRoot(routes),
-    MaterialModule
-  ],
   declarations: [
     AppComponent,
     CartComponent,
@@ -28,6 +30,23 @@ import { ProductComponent, ProductDetailsComponent } from './product';
     ProductComponent,
     ProductTileComponent,
     ProductDetailsComponent,
+  ],
+  imports: [
+    BrowserModule,
+    BrowserAnimationsModule,
+    FormsModule,
+    HttpModule,
+    HttpClientModule,
+    ReactiveFormsModule,
+    RouterModule.forRoot(routes),
+
+    MdButtonModule,
+    MdGridListModule,
+    MdSelectModule,
+    MdToolbarModule,
+    MdTabsModule,
+    MdIconModule,
+    MdInputModule
   ],
   providers: [
     CartResolver,
