@@ -1,2 +1,2 @@
-export * from './product.service';
+export { Product, ProductService } from './product.service';
 export * from './shopping-cart.service';
