@@ -1,28 +1,25 @@
 import { Injectable } from '@angular/core';
-import { Http } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/operator/map';
 
 @Injectable()
 export class ProductService {
 
-  constructor(private http: Http) {}
+  constructor(private http: HttpClient) {}
 
   getAll(): Observable<Product[]> {
-    return this.http.get('/data/products/all.json')
-      .map(resp => resp.json());
+    return this.http.get<Product[]>('/data/products/all.json');
   }
 
   getCategory(category: string): Observable<Product[]> {
-    return this.http.get(`/data/products/${category}.json`)
-      .map(resp => resp.json());
+    return this.http.get<Product[]>(`/data/products/${category}.json`);
   }
 
   getProductById(productId: string): Observable<Product> {
-    return this.http.get('/data/products/all.json')
-      .map(resp => resp.json().find(p => p.id === productId));
+    return this.http.get<Product[]>('/data/products/all.json')
+      .map(products => products.find(p => p.id === productId));
   }
-
 }
 
 export interface Product {
