@@ -1,4 +1,5 @@
 /* SystemJS module definition */
-declare var module: {
+declare var module: NodeModule;
+interface NodeModule {
   id: string;
-};
+}
